Rename interface to reduce code changes and delete metadata.go
