+++ pkg/controllers/csi/driver/volumes/host/publisher.go
@@ -34,7 +34,7 @@ import (
 
 const failedToGetOsAgentVolumePrefix = "failed to get OSMount from database: "
 
-func NewHostVolumePublisher(fs afero.Afero, mounter mount.Interface, db metadata.GormAccess, path metadata.PathResolver) csivolumes.Publisher {
+func NewHostVolumePublisher(fs afero.Afero, mounter mount.Interface, db metadata.Access, path metadata.PathResolver) csivolumes.Publisher {
 	return &HostVolumePublisher{
 		fs:      fs,
 		mounter: mounter,
@@ -46,7 +46,7 @@ func NewHostVolumePublisher(fs afero.Afero, mounter mount.Interface, db metadata
 type HostVolumePublisher struct {
 	fs      afero.Afero
 	mounter mount.Interface
-	db      metadata.GormAccess
+	db      metadata.Access
 	path    metadata.PathResolver
 }
